code cleanup 3

+++ Project 2.1/src/AI/MCTS/UCT.java
@@ -10,7 +10,7 @@ public class UCT {
         if (nodeVisit == 0) {
             return Integer.MAX_VALUE;
         }
-        return ((double) nodeWinScore / (double) nodeVisit) + 1.41 * Math.sqrt(Math.log(totalVisit) / (double) nodeVisit);
+        return (nodeWinScore / (double) nodeVisit) + 1.41 * Math.sqrt(Math.log(totalVisit) / (double) nodeVisit);
     }
 
     public static Node findBestChildWithUCT(Node node) {
